ready to prod step 3.
I give myself one last step to include breaking change then restman should be solidify for at least 1.x.x
bumping test to 75%
refacto/clean up structure
