+++ go.mod
@@ -1,8 +1,6 @@
 module github.com/philiphil/restman
 
-go 1.21.0
-
-toolchain go1.22.4
+go 1.23
 
 require (
 	github.com/gin-gonic/gin v1.10.0
