+++ orm/entity/model.go
@@ -2,5 +2,5 @@ package entity
 
 type Model[E IEntity] interface {
 	ToEntity() E
-	FromEntity(entity E) interface{}
+	FromEntity(entity E) any
 }
