+++ router/post.go
@@ -1,10 +1,9 @@
-package restman
+package router
 
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/philiphil/restman/errors"
 	method_type "github.com/philiphil/restman/method/MethodType"
-	"github.com/philiphil/restman/router"
 )
 
 func (r *ApiRouter[T]) Post(c *gin.Context) {
@@ -14,7 +13,7 @@ func (r *ApiRouter[T]) Post(c *gin.Context) {
 		return
 	}
 
-	if err := router.UnserializeBodyAndMerge(c, &entity); err != nil {
+	if err := UnserializeBodyAndMerge(c, &entity); err != nil {
 		c.AbortWithStatusJSON(err.(errors.ApiError).Code, err.(errors.ApiError).Message)
 		return
 	}
@@ -23,13 +22,13 @@ func (r *ApiRouter[T]) Post(c *gin.Context) {
 		c.AbortWithStatusJSON(errors.ErrDatabaseIssue.Code, errors.ErrDatabaseIssue.Message)
 		return
 	}
-	responseFormat, err := router.ParseAcceptHeader(c.GetHeader("Accept"))
+	responseFormat, err := ParseAcceptHeader(c.GetHeader("Accept"))
 	if err != nil {
 		c.AbortWithStatusJSON(err.(errors.ApiError).Code, err.(errors.ApiError).Message)
 		return
 	}
 
-	c.Render(201, router.SerializerRenderer{
+	c.Render(201, SerializerRenderer{
 		Data:   &entity,
 		Format: responseFormat,
 		Groups: r.GetMethodConfiguration(method_type.Get).SerializationGroups, //what is sent shall be compliant to get
