+++ router/security.go
@@ -1,4 +1,4 @@
-package restman
+package router
 
 import (
 	"github.com/gin-gonic/gin"
