+++ security/IRights.go
@@ -7,12 +7,12 @@ type WritingRights interface {
 	GetWritingRights() AuthorizationFunction
 }
 
-func HasReadingRights(obj interface{}) (ReadingRights, bool) {
+func HasReadingRights(obj any) (ReadingRights, bool) {
 	rr, ok := obj.(ReadingRights)
 	return rr, ok
 }
 
-func HasWritingRights(obj interface{}) (WritingRights, bool) {
+func HasWritingRights(obj any) (WritingRights, bool) {
 	rr, ok := obj.(WritingRights)
 	return rr, ok
 }
