+++ test/router/delete_test.go
@@ -1,4 +1,4 @@
-package restman_test
+package router_test
 
 import (
 	"fmt"
@@ -9,16 +9,14 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	. "github.com/philiphil/restman"
-	"github.com/philiphil/restman/format"
 	"github.com/philiphil/restman/method"
 	"github.com/philiphil/restman/orm"
 	"github.com/philiphil/restman/orm/entity"
 	"github.com/philiphil/restman/orm/repository"
-	"github.com/philiphil/restman/serializer"
+	. "github.com/philiphil/restman/router"
 )
 
-func TestApiRouter_Get(t *testing.T) {
+func TestApiRouter_delete(t *testing.T) {
 	getDB().AutoMigrate(&Test{})
 	getDB().Exec("DELETE FROM tests")
 	r := SetupRouter()
@@ -30,33 +28,36 @@ func TestApiRouter_Get(t *testing.T) {
 	)
 	test_.AllowRoutes(r)
 	context, _ := gin.CreateTestContext(httptest.NewRecorder())
-	test_.Get(context)
+	test_.Delete(context)
 
 	entity := Test{entity.Entity{Id: 1}}
 	repo.Create(&entity)
-
 	w := httptest.NewRecorder()
 
 	req, _ := http.NewRequest("GET", "/api/test/1", nil)
 	r.ServeHTTP(w, req)
-
 	if w.Code != http.StatusOK {
-		fmt.Println(w.Body.String())
 		t.Error("Failed to start server")
-	}
-	serializer := serializer.NewSerializer(format.JSON)
-	medium := Test{}
-	serializer.Deserialize(w.Body.String(), &medium)
-	if medium.Id != entity.Id {
-		t.Error("Expected entity with id " + fmt.Sprint(entity.Id) + " but got " + fmt.Sprint(medium.Id))
 	}
 
 	w = httptest.NewRecorder()
-	req, _ = http.NewRequest("GET", "/api/test/1", nil)
-	req.Header.Add("Accept", "Wrong")
+	req, _ = http.NewRequest("DELETE", "/api/test/1", nil)
 	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusNotAcceptable {
-		t.Error("should not accept")
+	if w.Code != http.StatusNoContent {
+		fmt.Println(w.Body.String())
+		fmt.Println(w.Code)
+		t.Error("Failed to delete")
 	}
+
+	if object, err := repo.GetByID(1); object != nil && err != nil {
+		t.Error("Failed to delete")
+	}
+
+	w = httptest.NewRecorder() //not sure why the recorder has to be reset sometimes
+	req, _ = http.NewRequest("DELETE", "/api/test/1", nil)
+	r.ServeHTTP(w, req)
+	if w.Code != http.StatusNotFound {
+		t.Error("Failed to delete")
+	}
+
 }
