+++ test/router/options_test.go
@@ -1,7 +1,7 @@
-package restman_test
+package router_test
 
 import (
-	"strconv"
+	"fmt"
 	"testing"
 
 	"net/http"
@@ -9,22 +9,20 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	. "github.com/philiphil/restman"
-	"github.com/philiphil/restman/format"
 	"github.com/philiphil/restman/method"
 	"github.com/philiphil/restman/orm"
 	"github.com/philiphil/restman/orm/entity"
 	"github.com/philiphil/restman/orm/repository"
-	"github.com/philiphil/restman/serializer"
+	"github.com/philiphil/restman/router"
 )
 
-func TestApiRouter_Head(t *testing.T) {
+func TestApiRouter_Options(t *testing.T) {
 	getDB().AutoMigrate(&Test{})
 	getDB().Exec("DELETE FROM tests")
 	r := SetupRouter()
 
 	repo := orm.NewORM[Test](repository.NewRepository[Test, Test](getDB()))
-	test_ := NewApiRouter[Test](
+	test_ := router.NewApiRouter[Test](
 		*repo,
 		method.DefaultApiMethods(),
 	)
@@ -44,23 +42,15 @@ func TestApiRouter_Head(t *testing.T) {
 	entity := Test{entity.Entity{Id: 1}}
 	repo.Create(&entity)
 	//serializer mon truc en json et compare avec le json de la reponse
-	req, _ = http.NewRequest("HEAD", "/api/test/1", nil)
+	req, _ = http.NewRequest("OPTIONS", "/api/test/1", nil)
 	r.ServeHTTP(w, req)
-	serializer := serializer.NewSerializer(format.JSON)
-	json, _ := serializer.Serialize(entity)
-	if w.Header().Get("Content-Type") != "application/json" {
-		t.Error("Content-Type should be application/json")
-	}
-	if w.Header().Get("Content-Length") != strconv.Itoa(len(json)) {
-		t.Error("Content-Length should be equal to the length of the json")
-	}
 
-	w = httptest.NewRecorder()
-	req, _ = http.NewRequest("HEAD", "/api/test/1", nil)
-	req.Header.Set("Accept", "cac/esfes")
-	r.ServeHTTP(w, req)
-	if w.Code != http.StatusNotAcceptable {
-		t.Error("should be not acceptable")
+	if w.Header().Get("Allow") != "GET,POST,PUT,PATCH,DELETE,HEAD,OPTIONS" {
+		fmt.Println(w.Header().Get("Allow"))
+		t.Error("Allow should be GET,POST,PUT,PATCH,DELETE,HEAD,OPTIONS")
+	}
+	if w.Header().Get("Content-Length") != "0" {
+		t.Error("Content-Length should be 0")
 	}
 
 }
