+++ test/router/put_test.go
@@ -1,4 +1,4 @@
-package restman_test
+package router_test
 
 import (
 	"bytes"
@@ -6,14 +6,13 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	. "github.com/philiphil/restman"
 	"github.com/philiphil/restman/method"
 	"github.com/philiphil/restman/orm"
-	"github.com/philiphil/restman/orm/entity"
 	"github.com/philiphil/restman/orm/repository"
+	. "github.com/philiphil/restman/router"
 )
 
-func TestApiRouter_patch(t *testing.T) {
+func TestApiRouter_put(t *testing.T) {
 	getDB().AutoMigrate(&Test{})
 	getDB().Exec("DELETE FROM tests")
 	r := SetupRouter()
@@ -27,22 +26,19 @@ func TestApiRouter_patch(t *testing.T) {
 
 	w := httptest.NewRecorder()
 
-	req, _ := http.NewRequest("PATCH", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test"}`)))
-	req.Header.Set("Accept", "application/json")
-	req.Header.Set("Content-Type", "application/json")
-	r.ServeHTTP(w, req)
-	if w.Code != http.StatusNotFound {
-		t.Error("should be no content")
-	}
-	getDB().Create(&Test{entity.Entity{Id: 2, Name: "test"}})
-	w = httptest.NewRecorder()
-	req, _ = http.NewRequest("PATCH", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test2"}`)))
+	req, _ := http.NewRequest("PUT", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test"}`)))
 	req.Header.Set("Accept", "application/json")
 	req.Header.Set("Content-Type", "application/json")
 	r.ServeHTTP(w, req)
 	if w.Code != http.StatusNoContent {
 		t.Error("should be no content")
 	}
+	w = httptest.NewRecorder()
+	req, _ = http.NewRequest("PUT", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test2"}`)))
+	req.Header.Set("Accept", "application/json")
+	req.Header.Set("Content-Type", "application/json")
+	r.ServeHTTP(w, req)
+
 	var medium Test
 	getDB().First(&medium, 2)
 	if medium.Name != "test2" {
@@ -50,7 +46,7 @@ func TestApiRouter_patch(t *testing.T) {
 	}
 	//faillures
 	w = httptest.NewRecorder()
-	req, _ = http.NewRequest("PATCH", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test2"}`)))
+	req, _ = http.NewRequest("PUT", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test2"}`)))
 	req.Header.Set("Accept", "application/json")
 	req.Header.Set("Content-Type", "afesfs")
 	r.ServeHTTP(w, req)
@@ -58,7 +54,7 @@ func TestApiRouter_patch(t *testing.T) {
 		t.Error("should be bad request")
 	}
 	w = httptest.NewRecorder()
-	req, _ = http.NewRequest("PATCH", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test2"`)))
+	req, _ = http.NewRequest("PUT", "/api/test/2", bytes.NewBuffer([]byte(`{"name":"test2"`)))
 	req.Header.Set("Accept", "application/json")
 	req.Header.Set("Content-Type", "application/json")
 	r.ServeHTTP(w, req)
